+++ src/edge.ts
@@ -67,6 +67,8 @@ export interface Edge {
     type?: any;
     /** List of markers to draw for this edge. */
     markers?: Marker[];
+    /** Markers to draw at the start of this edge. */
+    markerStart?: Marker;
     /** Markers to draw at the end of this edge. */
     markerEnd?: Marker;
     /** List of text components of this edge. */